The deleteIDColumnIfNecessary method is executed only in dual write mode

+++ iotdb-2.0/src/main/java/cn/edu/tsinghua/iot/benchmark/iotdb200/DMLStrategy/SessionStrategy.java
@@ -155,7 +155,9 @@ public class SessionStrategy extends DMLStrategy {
     List<IMeasurementSchema> schemaList = new ArrayList<>();
     List<Tablet.ColumnType> columnTypes = new ArrayList<>();
     List<Sensor> sensors = batch.getDeviceSchema().getSensors();
-    iotdb.deleteIDColumnIfNecessary(columnTypes, sensors, batch);
+    if (config.isIS_DOUBLE_WRITE()) {
+      iotdb.deleteIDColumnIfNecessary(columnTypes, sensors, batch);
+    }
     iotdb.addIDColumnIfNecessary(columnTypes, sensors, batch);
     int sensorIndex = 0;
     for (Sensor sensor : sensors) {
